# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Instrument API",
    "description": "API documentation for instrument and RFID tracking system",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/LF": {
      "post": {
        "summary": "Insert or update LF RFID",
        "tags": [
          "RFID"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "LF": {
                    "type": "string"
                  },
                  "lf": {
                    "type": "string"
                  },
                  "HF": {
                    "type": "string"
                  },
                  "hf": {
                    "type": "string"
                  },
                  "rfid": {
                    "type": "string"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {}
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {}
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/HF": {
      "post": {
        "summary": "Insert or update HF RFID",
        "tags": [
          "RFID"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "LF": {
                    "type": "string"
                  },
                  "lf": {
                    "type": "string"
                  },
                  "HF": {
                    "type": "string"
                  },
                  "hf": {
                    "type": "string"
                  },
                  "rfid": {
                    "type": "string"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {}
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {}
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/LF/check": {
      "post": {
        "summary": "Check LF RFID existence and association",
        "tags": [
          "RFID"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "LF": {
                    "type": "string"
                  },
                  "lf": {
                    "type": "string"
                  },
                  "HF": {
                    "type": "string"
                  },
                  "hf": {
                    "type": "string"
                  },
                  "rfid": {
                    "type": "string"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "checked_rfid": {
                      "type": "string",
                      "nullable": true
                    },
                    "type": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message",
                    "checked_rfid"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "checked_rfid": {
                      "enum": [
                        "null"
                      ],
                      "nullable": true
                    }
                  },
                  "required": [
                    "message",
                    "checked_rfid"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "checked_rfid": {
                      "enum": [
                        "null"
                      ],
                      "nullable": true
                    }
                  },
                  "required": [
                    "message",
                    "checked_rfid"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "checked_rfid": {
                      "enum": [
                        "null"
                      ],
                      "nullable": true
                    }
                  },
                  "required": [
                    "message",
                    "checked_rfid"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/HF/check": {
      "post": {
        "summary": "Check HF RFID existence and association",
        "tags": [
          "RFID"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "LF": {
                    "type": "string"
                  },
                  "lf": {
                    "type": "string"
                  },
                  "HF": {
                    "type": "string"
                  },
                  "hf": {
                    "type": "string"
                  },
                  "rfid": {
                    "type": "string"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "checked_rfid": {
                      "type": "string",
                      "nullable": true
                    },
                    "type": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message",
                    "checked_rfid"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "checked_rfid": {
                      "enum": [
                        "null"
                      ],
                      "nullable": true
                    }
                  },
                  "required": [
                    "message",
                    "checked_rfid"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "checked_rfid": {
                      "enum": [
                        "null"
                      ],
                      "nullable": true
                    }
                  },
                  "required": [
                    "message",
                    "checked_rfid"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "checked_rfid": {
                      "enum": [
                        "null"
                      ],
                      "nullable": true
                    }
                  },
                  "required": [
                    "message",
                    "checked_rfid"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/LF/load": {
      "get": {
        "summary": "Load LF RFID records updated after timestamp",
        "tags": [
          "RFID"
        ],
        "parameters": [
          {
            "schema": {
              "type": "number",
              "default": 0
            },
            "in": "query",
            "name": "timestamp",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ids": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "data": {
                      "type": "array",
                      "items": {}
                    }
                  },
                  "required": [
                    "ids",
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {}
                    }
                  },
                  "required": [
                    "message",
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {}
                    }
                  },
                  "required": [
                    "message",
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/HF/load": {
      "get": {
        "summary": "Load HF RFID records updated after timestamp",
        "tags": [
          "RFID"
        ],
        "parameters": [
          {
            "schema": {
              "type": "number",
              "default": 0
            },
            "in": "query",
            "name": "timestamp",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ids": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "data": {
                      "type": "array",
                      "items": {}
                    }
                  },
                  "required": [
                    "ids",
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {}
                    }
                  },
                  "required": [
                    "message",
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {}
                    }
                  },
                  "required": [
                    "message",
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/LF/load-deleted": {
      "get": {
        "summary": "Load deleted LF RFID records after timestamp",
        "tags": [
          "RFID"
        ],
        "parameters": [
          {
            "schema": {
              "type": "number",
              "default": 0
            },
            "in": "query",
            "name": "timestamp",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ids": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "data": {
                      "type": "array",
                      "items": {}
                    }
                  },
                  "required": [
                    "ids",
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {}
                    }
                  },
                  "required": [
                    "message",
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {}
                    }
                  },
                  "required": [
                    "message",
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": []
}